Created dashboard.active and dashboard.dormant packages
