+++ app/src/main/java/io/lunarlogic/aircasting/screens/dashboard/DashboardPagerAdapter.kt
@@ -7,8 +7,8 @@ import androidx.fragment.app.FragmentPagerAdapter
 import io.lunarlogic.aircasting.R
 import io.lunarlogic.aircasting.screens.dashboard.fixed.FixedFragment
 import io.lunarlogic.aircasting.screens.dashboard.following.FollowingFragment
-import io.lunarlogic.aircasting.screens.dashboard.mobile.MobileActiveFragment
-import io.lunarlogic.aircasting.screens.dashboard.mobile.MobileDormantFragment
+import io.lunarlogic.aircasting.screens.dashboard.active.MobileActiveFragment
+import io.lunarlogic.aircasting.screens.dashboard.dormant.MobileDormantFragment
 import io.lunarlogic.aircasting.models.Session
 
 enum class SessionsTab(val value: Int){
