+++ app/src/main/java/io/lunarlogic/aircasting/screens/dashboard/active/MobileActiveController.kt
@@ -1,18 +1,16 @@
-package io.lunarlogic.aircasting.screens.dashboard.mobile
+package io.lunarlogic.aircasting.screens.dashboard.active
 
 import androidx.fragment.app.FragmentActivity
 import androidx.lifecycle.LifecycleOwner
-import androidx.lifecycle.LiveData
-import io.lunarlogic.aircasting.database.data_classes.SessionWithStreamsAndMeasurementsDBObject
 import io.lunarlogic.aircasting.events.StopRecordingEvent
 import io.lunarlogic.aircasting.lib.NavigationController
 import io.lunarlogic.aircasting.lib.Settings
 import io.lunarlogic.aircasting.models.ActiveSessionsObserver
+import io.lunarlogic.aircasting.models.Session
+import io.lunarlogic.aircasting.models.SessionsViewModel
 import io.lunarlogic.aircasting.screens.dashboard.DashboardPagerAdapter
 import io.lunarlogic.aircasting.screens.dashboard.SessionsController
-import io.lunarlogic.aircasting.models.SessionsViewModel
 import io.lunarlogic.aircasting.screens.dashboard.SessionsViewMvc
-import io.lunarlogic.aircasting.models.Session
 import org.greenrobot.eventbus.EventBus
 
 class MobileActiveController(
@@ -24,7 +22,8 @@ class MobileActiveController(
 ): SessionsController(mRootActivity, mViewMvc, mSessionsViewModel, mLifecycleOwner, mSettings),
     SessionsViewMvc.Listener {
 
-    private var mSessionsObserver = ActiveSessionsObserver(mLifecycleOwner, mSessionsViewModel, mViewMvc)
+    private var mSessionsObserver =
+        ActiveSessionsObserver(mLifecycleOwner, mSessionsViewModel, mViewMvc)
 
     override fun registerSessionsObserver() {
         mSessionsObserver.observe(mSessionsViewModel.loadMobileActiveSessionsWithMeasurements())
@@ -42,7 +41,10 @@ class MobileActiveController(
         val event = StopRecordingEvent(sessionUUID)
         EventBus.getDefault().post(event)
 
-        val tabId = DashboardPagerAdapter.tabIndexForSessionType(Session.Type.MOBILE, Session.Status.FINISHED)
+        val tabId = DashboardPagerAdapter.tabIndexForSessionType(
+            Session.Type.MOBILE,
+            Session.Status.FINISHED
+        )
         NavigationController.goToDashboard(tabId)
     }
