+++ app/src/main/java/io/lunarlogic/aircasting/screens/dashboard/active/MobileActiveRecyclerAdapter.kt
@@ -1,10 +1,9 @@
-package io.lunarlogic.aircasting.screens.dashboard.mobile
+package io.lunarlogic.aircasting.screens.dashboard.active
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.fragment.app.FragmentManager
 import io.lunarlogic.aircasting.screens.dashboard.SessionsRecyclerAdapter
-
 
 class MobileActiveRecyclerAdapter(
     private val mInflater: LayoutInflater,
