+++ app/src/main/java/io/lunarlogic/aircasting/screens/dashboard/active/MobileActiveViewMvcImpl.kt
@@ -1,34 +1,33 @@
-package io.lunarlogic.aircasting.screens.dashboard.mobile
+package io.lunarlogic.aircasting.screens.dashboard.active
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.fragment.app.FragmentManager
+import io.lunarlogic.aircasting.models.Session
 import io.lunarlogic.aircasting.screens.dashboard.SessionsRecyclerAdapter
 import io.lunarlogic.aircasting.screens.dashboard.SessionsViewMvcImpl
-import io.lunarlogic.aircasting.models.Session
 
-
-class MobileDormantViewMvcImpl(
+class MobileActiveViewMvcImpl(
     inflater: LayoutInflater,
     parent: ViewGroup?,
     supportFragmentManager: FragmentManager
-): SessionsViewMvcImpl<MobileDormantSessionViewMvc.Listener>(inflater, parent, supportFragmentManager),
-    MobileDormantSessionViewMvc.Listener {
+): SessionsViewMvcImpl<MobileActiveSessionViewMvc.Listener>(inflater, parent, supportFragmentManager),
+    MobileActiveSessionViewMvc.Listener {
 
     override fun buildAdapter(
         inflater: LayoutInflater,
         supportFragmentManager: FragmentManager
-    ): SessionsRecyclerAdapter<MobileDormantSessionViewMvc.Listener> {
-        return MobileDormantRecyclerAdapter(
+    ): SessionsRecyclerAdapter<MobileActiveSessionViewMvc.Listener> {
+        return MobileActiveRecyclerAdapter(
             inflater,
             this,
             supportFragmentManager
         )
     }
 
-    override fun onSessionDeleteClicked(session: Session) {
+    override fun onSessionStopClicked(session: Session) {
         for (listener in listeners) {
-            listener.onDeleteSessionClicked(session.uuid)
+            listener.onStopSessionClicked(session.uuid)
         }
     }
 }
