+++ app/src/main/java/io/lunarlogic/aircasting/screens/dashboard/dormant/MobileDormantController.kt
@@ -1,20 +1,14 @@
-package io.lunarlogic.aircasting.screens.dashboard.mobile
+package io.lunarlogic.aircasting.screens.dashboard.dormant
 
 import androidx.fragment.app.FragmentActivity
 import androidx.lifecycle.LifecycleOwner
-import androidx.lifecycle.LiveData
-import io.lunarlogic.aircasting.database.DatabaseProvider
-import io.lunarlogic.aircasting.database.data_classes.SessionWithStreamsAndMeasurementsDBObject
 import io.lunarlogic.aircasting.events.DeleteSessionEvent
-import io.lunarlogic.aircasting.exceptions.ErrorHandler
 import io.lunarlogic.aircasting.lib.Settings
-import io.lunarlogic.aircasting.models.ActiveSessionsObserver
 import io.lunarlogic.aircasting.models.DormantSessionsObserver
 import io.lunarlogic.aircasting.screens.dashboard.SessionsController
 import io.lunarlogic.aircasting.models.SessionsViewModel
 import io.lunarlogic.aircasting.screens.dashboard.SessionsViewMvc
 import io.lunarlogic.aircasting.models.Session
-import io.lunarlogic.aircasting.networking.services.DownloadMeasurementsService
 import org.greenrobot.eventbus.EventBus
 
 class MobileDormantController(
