+++ app/src/main/java/io/lunarlogic/aircasting/screens/dashboard/dormant/MobileDormantSessionViewMvc.kt
@@ -1,12 +1,12 @@
-package io.lunarlogic.aircasting.screens.dashboard.mobile
+package io.lunarlogic.aircasting.screens.dashboard.dormant
 
 import io.lunarlogic.aircasting.screens.dashboard.SessionCardListener
 import io.lunarlogic.aircasting.screens.dashboard.SessionViewMvc
 import io.lunarlogic.aircasting.models.Session
 
-interface MobileActiveSessionViewMvc:
-    SessionViewMvc<MobileActiveSessionViewMvc.Listener> {
+interface MobileDormantSessionViewMvc:
+    SessionViewMvc<MobileDormantSessionViewMvc.Listener> {
     interface Listener: SessionCardListener {
-        fun onSessionStopClicked(session: Session)
+        fun onSessionDeleteClicked(session: Session)
     }
 }
