--- app/src/main/java/io/lunarlogic/aircasting/screens/dashboard/mobile/MobileDormantSessionViewMvc.kt
@@ -1,12 +1,0 @@
-package io.lunarlogic.aircasting.screens.dashboard.mobile
-
-import io.lunarlogic.aircasting.screens.dashboard.SessionCardListener
-import io.lunarlogic.aircasting.screens.dashboard.SessionViewMvc
-import io.lunarlogic.aircasting.models.Session
-
-interface MobileDormantSessionViewMvc:
-    SessionViewMvc<MobileDormantSessionViewMvc.Listener> {
-    interface Listener: SessionCardListener {
-        fun onSessionDeleteClicked(session: Session)
-    }
-}
